requirements.txt
datetime
pydantic
redis
asyncio
websockets
torch
pillow
websocket
torchvision
numpy
python-dotenv
prometheus_client
uvicorn
pydantic-settings
fastapi
requests